.history/requirements_20250328150643.txt
streamlit==1.29.0
python-docx==1.1.0
plotly==5.18.0
wordcloud==1.9.3
pandas==2.2.1
typing-extensions==4.9.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0